Split env detection to separate file
